+++ NEWS.md
@@ -5,6 +5,12 @@ the EndBASIC language definition nor the API exposed by this crate.  Expect
 them to change at any time (especially the Rust API).  Version numbers will
 not adhere to semantic versioning until 1.0.0.**
 
+## Changes in version X.Y.Z
+
+**STILL UNDER DEVELOPMENT; NOT RELEASED YET.**
+
+*   Added the `CLS` command for terminal manipulation.
+
 ## Changes in version 0.2.0
 
 **Released on 2020-05-07.**